report: show in Viewer/Browser + tempfile() by default

+++ R/report_RLum.R
@@ -23,14 +23,19 @@
 #' @param object (\bold{required}): 
 #' The object to be reported on, preferably of any \code{RLum}-class.
 #' 
-#' @param file \code{\link{character}} (\bold{required}): 
-#' A character string naming the output file.
+#' @param file \code{\link{character}} (\bold{with default}): 
+#' A character string naming the output file. If no filename is provided a 
+#' temporary file is created.
 #' 
 #' @param title \code{\link{character}} (with default):
 #' A character string specifying the title of the document.
 #' 
 #' @param timestamp \code{\link{logical}} (with default):
 #' \code{TRUE} to add a timestamp to the filename (suffix).
+#' 
+#' @param launch.browser \code{\link{logical}} (with default):
+#' \code{TRUE} to open the HTML file in the system's default web browser after
+#' it has been rendered.
 #' 
 #' @param clean \code{\link{logical}} (with default): 
 #' \code{TRUE} to clean intermediate files created during rendering.
@@ -96,14 +101,15 @@
 #' report_RLum(object = x, file = "~/arbitray_list")
 #' }
 report_RLum <- function(object, 
-                        file,
+                        file  = tempfile(),
                         title = "RLum.Report",
-                        timestamp = TRUE, 
+                        timestamp = TRUE,
+                        launch.browser = FALSE,
                         clean = TRUE, 
                         ...) {
   
   ## INPUT VALIDATION ----
-  
+ 
   # check if required namespace(s) are available
   if (!requireNamespace("rmarkdown", quietly = TRUE))
     stop("Creating object reports requires the 'rmarkdown' package.",
@@ -114,11 +120,9 @@ report_RLum <- function(object,
          " To install this package run 'install.packages('pander')' in your R console.", 
          call. = FALSE)
   
-  # has a filename been provided?
-  if (missing(file))
-    stop("Please provide a directory/filename.", call. = FALSE)
-  
   # CREATE FILE
+  isTemp <- missing(file)
+  
   if (!grepl(".rmd$", file, ignore.case = TRUE))
     file <- paste0(file, ".Rmd")
   
@@ -132,6 +136,9 @@ report_RLum <- function(object,
                  ignore.case = TRUE)
   
   file.create(file)
+  
+  # sanitize file name
+  file <- gsub("\\\\", "\\/", file)
   
   # OPEN CONNECTION
   tmp <- file(file, open = "w")
@@ -272,7 +279,7 @@ report_RLum <- function(object,
   
   # PLOTTING
   if (length(grep("RLum", class(object)))) {
-    
+
     writeLines(paste("\n\n<hr># Plots\n\n"), tmp)
     writeLines(paste0(
       "```{r}\n",
@@ -296,7 +303,7 @@ report_RLum <- function(object,
       if (object@originator %in% models) {
         writeLines(paste0(
           "```{r}\n",
-          "plot_AbanicoPlot(get_RLum(x, 'data')) \n",
+          "plot_AbanicoPlot(x) \n",
           "plot_Histogram(x) \n",
           "plot_KDE(x) \n", 
           "```"),
@@ -311,8 +318,24 @@ report_RLum <- function(object,
   on.exit(closeAllConnections())
   rmarkdown::render(file, clean = clean)
   
+  # SHOW FILE
+  file.html <- gsub(".rmd$", ".html", file, ignore.case = TRUE)
+  
+  if (isTemp) {
+    try(rstudioapi::viewer(file.html))
+  } else {
+    # The Viewer Pane only works for files in a sessions temp directory
+    # see: https://support.rstudio.com/hc/en-us/articles/202133558-Extending-RStudio-with-the-Viewer-Pane
+    file.copy(file.html, file.path(tempdir(), "report.html"))
+    try(rstudioapi::viewer(file.path(tempdir(), "report.html")))
+  }
+    
+  if (launch.browser)
+    try(browseURL(file.html))
+  
   # CLEANUP
-  file.remove(file)
+  if (clean)
+    file.remove(file)
 }
